move ProxyBasicRule
