+++ sharding-proxy/src/main/java/io/shardingsphere/proxy/config/RuleRegistry.java
@@ -18,6 +18,7 @@
 package io.shardingsphere.proxy.config;
 
 import com.google.common.eventbus.Subscribe;
+import io.shardingsphere.core.api.config.ProxyBasicRule;
 import io.shardingsphere.core.api.config.ShardingRuleConfiguration;
 import io.shardingsphere.core.constant.ConnectionMode;
 import io.shardingsphere.core.constant.DatabaseType;
@@ -26,7 +27,6 @@ import io.shardingsphere.core.constant.properties.ShardingPropertiesConstant;
 import io.shardingsphere.core.constant.transaction.TransactionType;
 import io.shardingsphere.core.executor.ShardingExecuteEngine;
 import io.shardingsphere.core.metadata.ShardingMetaData;
-import io.shardingsphere.core.orche.config.ProxyBasicRule;
 import io.shardingsphere.core.event.orche.config.ProxyConfigurationEventBusEvent;
 import io.shardingsphere.core.event.orche.state.CircuitStateEventBusEvent;
 import io.shardingsphere.core.event.orche.state.DisabledStateEventBusEvent;
